Disabled JS scrolling on the consolidated dashboard + analyze page
