+++ utils/referrer-link-util.ts
@@ -1,4 +1,4 @@
-const safeReferrerLinkRegex = /^(\/[a-zA-Z0-9?=]+)+$/
+const safeReferrerLinkRegex = /^(\/[a-zA-Z0-9?=#]+)+$/
 
 export const isSafeReferrerLink = (referrerLink: string) => {
   return safeReferrerLinkRegex.test(referrerLink);
